🔢 create input component
